refactor: external pages
